**Rename loginpage and mainpage component** Fixes #14
