+++ WhereSweat/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CategoriesPageComponent } from './components/categories-page/categories-page.component';
 import { CategoryPageComponent } from './components/category-page/category-page.component';
-import { LoginpageComponent } from './components/loginpage/loginpage.component';
-import { MainpageComponent } from './components/mainpage/mainpage.component';
+import { LoginpageComponent } from './components/login-page/login-page.component';
+import { MainpageComponent } from './components/main-page/main-page.component';
 import { RegistryPageComponent } from './components/registry-page/registry-page.component';
 import { ProfileSettingsComponent } from './components/profile-settings/profile-settings.component';
