+++ WhereSweat/src/app/app.module.ts
@@ -3,8 +3,8 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LoginpageComponent } from './components/loginpage/loginpage.component';
-import { MainpageComponent } from './components/mainpage/mainpage.component';
+import { LoginpageComponent } from './components/login-page/login-page.component';
+import { MainpageComponent } from './components/main-page/main-page.component';
 import { CategoriesPageComponent } from './components/categories-page/categories-page.component';
 import { CategoryPageComponent } from './components/category-page/category-page.component';
 import { RegistryPageComponent } from './components/registry-page/registry-page.component';
