+++ WhereSweat/src/app/components/main-page/main-page.component.ts
